Moved print_dups() to its own module
